fix(hipsparse): support ell_nnz beyond INT32_MAX in the hybmv/csr2hyb host path (AISPARSE-681)
Review feedback (jsandham): hybmv should support matrices whose ELL part
exceeds INT32_MAX elements while m stays an int, so hold ell_nnz as an
int64_t instead of casting dhyb->ell_nnz down to int.

host_hybmv now takes int64_t ell_nnz and computes the ELL element index
ELL_IND(i, p, m, ell_width) = p * m + i in 64-bit. The 32-bit index was
the same overflow this PR fixes on the device side: with ell_nnz past
2^31 the host reference read the wrong ELL slot (or out of bounds) and
the comparison against the GPU result became meaningless.

testing_csr2hyb's host conversion had the same truncation: ell_nnz is
int64_t but was assigned the 32-bit product ell_width * m. Compute it in
64-bit so the host gold ELL buffers are sized correctly.

+++ projects/hipsparse/clients/include/testing_csr2hyb.hpp
@@ -257,7 +257,7 @@ void testing_csr2hyb(Arguments argus)
             ell_width = user_ell_width;
         }
 
-        ell_nnz = ell_width * m;
+        ell_nnz = static_cast<int64_t>(ell_width) * m;
 
         // Determine COO nnz
         for(int i = 0; i < m; ++i)
@@ -277,7 +277,7 @@ void testing_csr2hyb(Arguments argus)
             int row_nnz = hcsr_row_ptr[i + 1] - hcsr_row_ptr[i];
             ell_width   = (row_nnz > ell_width) ? row_nnz : ell_width;
         }
-        ell_nnz = ell_width * m;
+        ell_nnz = static_cast<int64_t>(ell_width) * m;
     }
 
     // Allocate host memory
